+++ packages/editor/src/store/actions.js
@@ -10,6 +10,7 @@ import deprecated from '@wordpress/deprecated';
 import { controls } from '@wordpress/data';
 import { apiFetch } from '@wordpress/data-controls';
 import { parse, synchronizeBlocksWithTemplate } from '@wordpress/blocks';
+import { store as noticesStore } from '@wordpress/notices';
 
 /**
  * Internal dependencies
@@ -74,6 +75,27 @@ export function* setupEditor( post, edits, template ) {
 	) {
 		yield editPost( edits );
 	}
+}
+
+/**
+ * Initiliazes an FSE template into the core-data store.
+ * We could avoid this action entirely by having a fallback if the edit is undefined.
+ *
+ * @param {Object} template  Template object.
+ */
+export function* __unstableSetupTemplate( template ) {
+	const blocks = parse( template.content.raw );
+	yield controls.dispatch(
+		'core',
+		'editEntityRecord',
+		'postType',
+		template.type,
+		template.id,
+		{
+			blocks,
+		},
+		{ __unstableShouldCreateUndoLevel: false }
+	);
 }
 
 /**
@@ -278,7 +300,7 @@ export function* savePost( options = {} ) {
 		} );
 		if ( args.length ) {
 			yield controls.dispatch(
-				'core/notices',
+				noticesStore,
 				'createErrorNotice',
 				...args
 			);
@@ -300,7 +322,7 @@ export function* savePost( options = {} ) {
 		} );
 		if ( args.length ) {
 			yield controls.dispatch(
-				'core/notices',
+				noticesStore,
 				'createSuccessNotice',
 				...args
 			);
@@ -354,7 +376,7 @@ export function* trashPost() {
 		postTypeSlug
 	);
 	yield controls.dispatch(
-		'core/notices',
+		noticesStore,
 		'removeNotice',
 		TRASH_POST_NOTICE_ID
 	);
@@ -368,7 +390,7 @@ export function* trashPost() {
 		yield controls.dispatch( STORE_NAME, 'savePost' );
 	} catch ( error ) {
 		yield controls.dispatch(
-			'core/notices',
+			noticesStore,
 			'createErrorNotice',
 			...getNotificationArgumentsForTrashFail( { error } )
 		);